+++ src/main/Game.java
@@ -1,6 +1,9 @@
 package main;
 
+import java.awt.Graphics;
 import java.awt.Toolkit;
+
+import entities.Player;
 
 public class Game implements Runnable {
 
@@ -10,13 +13,20 @@ public class Game implements Runnable {
 	private final int FPS_SET = 120;
 	private final int UPS_SET = 200;
 
+	private Player player;
+
 	public Game() {
 
-		gamePanel = new GamePanel();
+		initClasses();
+		gamePanel = new GamePanel(this);
 		gameWindow = new GameWindow(gamePanel);
 		gamePanel.requestFocus();
-		startGameLoop();
+		
+		startGameLoop(); // Alaways last line of the constructor
+	}
 
+	private void initClasses() {
+		player = new Player(200, 200);
 	}
 
 	private void startGameLoop() {
@@ -25,7 +35,11 @@ public class Game implements Runnable {
 	}
 
 	private void update() {
-		gamePanel.updateGame();
+		player.update();
+	}
+
+	public void render(Graphics g) {
+		player.render(g);
 	}
 
 	@Override
@@ -72,6 +86,14 @@ public class Game implements Runnable {
 			}
 		}
 
+	}
+
+	public void windowFocusLost() {
+		player.resetDirectionBooleans();
+	}
+
+	public Player getPlayer() {
+		return player;
 	}
 
 }