+++ src/main/GameWindow.java
@@ -1,5 +1,8 @@
 package main;
 
+import java.awt.event.WindowEvent;
+import java.awt.event.WindowFocusListener;
+
 import javax.swing.JFrame;
 
 public class GameWindow {
@@ -15,6 +18,18 @@ public class GameWindow {
 		jframe.setResizable(false);
 		jframe.pack();
 		jframe.setVisible(true);
+		jframe.addWindowFocusListener(new WindowFocusListener() {
+
+			@Override
+			public void windowGainedFocus(WindowEvent arg0) {
+			}
+
+			@Override
+			public void windowLostFocus(WindowEvent arg0) {
+				gamePanel.getGame().windowFocusLost();
+			}
+			
+		});
 
 	}